Display ImageGrid in Browse page
